+++ src/main/java/io/github/tlsdla1235/seniormealplan/domain/recipe/Recipe.java
@@ -18,7 +18,7 @@ public class Recipe {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "recipe_id")
-    private Integer recipeId;
+    private Long recipeId;
 
     @Column(nullable = false)
     private String name;
