+++ src/main/java/io/github/tlsdla1235/seniormealplan/domain/recipe/UserBookmarkedRecipe.java
@@ -16,7 +16,7 @@ public class UserBookmarkedRecipe {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "bookmark_id")
-    private Integer bookmarkId;
+    private Long bookmarkId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
